Move TransformResult to its own file
